Canonicalize KDA host weight schema
