+++ models/demos/deepseek_v3_d_p/reference/kda/weights.py
@@ -29,6 +29,15 @@ KDA_LOW_RANK_GATE_WEIGHT_NAMES = ("g_a_proj.weight", "g_b_proj.weight")
 KDA_FULL_RANK_GATE_WEIGHT_NAMES = ("g_proj.weight",)
 
 
+def normalize_kda_a_log(a_log: torch.Tensor, config: KDAConfig) -> torch.Tensor:
+    """Normalize checkpoint head padding into canonical ``[1,1,H,1]`` form."""
+    if a_log.numel() == config.num_heads:
+        return a_log.reshape(1, 1, config.num_heads, 1)
+    if config.num_heads == 96 and a_log.numel() == 128:
+        return a_log.reshape(-1)[: config.num_heads].reshape(1, 1, config.num_heads, 1)
+    raise ValueError(f"A_log has {a_log.numel()} entries; expected {config.num_heads} logical heads")
+
+
 def required_kda_weight_names(config: KDAConfig) -> tuple[str, ...]:
     """Return the canonical layer-local checkpoint keys for ``config``."""
     gate_names = KDA_FULL_RANK_GATE_WEIGHT_NAMES if config.use_full_rank_gate else KDA_LOW_RANK_GATE_WEIGHT_NAMES
@@ -73,3 +82,14 @@ def validate_kda_weights(weights: Mapping[str, torch.Tensor], config: KDAConfig)
             raise ValueError(f"missing KDA weight: {name}") from error
         if tuple(weight.shape) != expected_shape:
             raise ValueError(f"{name} shape {tuple(weight.shape)} != {expected_shape}")
+
+
+def normalize_kda_state_dict(state_dict: Mapping[str, torch.Tensor], config: KDAConfig) -> dict[str, torch.Tensor]:
+    """Copy, canonicalize, and validate a layer-local host-weight mapping."""
+    normalized = dict(state_dict)
+    try:
+        normalized["A_log"] = normalize_kda_a_log(normalized["A_log"], config)
+    except KeyError as error:
+        raise ValueError("missing KDA weight: A_log") from error
+    validate_kda_weights(normalized, config)
+    return normalized
