+++ models/demos/deepseek_v3_d_p/tests/kda/checkpoint_utils.py
@@ -13,8 +13,7 @@ import torch
 from safetensors import safe_open
 
 from models.demos.deepseek_v3_d_p.reference.kda.config import KDAConfig
-from models.demos.deepseek_v3_d_p.reference.kda.weights import required_kda_weight_names
-from models.demos.deepseek_v3_d_p.tt.kda.weight_schema import normalize_kda_state_dict
+from models.demos.deepseek_v3_d_p.reference.kda.weights import normalize_kda_state_dict, required_kda_weight_names
 
 KIMI_K3_FIRST_KDA_LAYER = 1
 KIMI_K3_HF_REVISION = "9f62e4e9fffbd0a83ddd60e1c209d828994b3569"
